added a py wrapper

+++ lib/libast.h
@@ -31,7 +31,6 @@
 #ifndef _LIBAST_H_
 #define _LIBAST_H_
 
-#include <stdio.h>
 #include <stdbool.h>
 
 /*============================================================================*\
@@ -133,7 +132,7 @@ Arguments:
 Return:
   Zero on success; non-zero on error.
 ******************************************************************************/
-int ast_eval_num(ast_t *ast, void *value, const void *var, const long size);
+int ast_eval_num(ast_t *ast, void *value, const void *var, const size_t size);
 
 /******************************************************************************
 Function `ast_perror`:
